apps/holiday-planner-api/holiday_planner/requirements.txt
asgiref==3.7.2
cffi==1.15.1
Django==4.2.6
django-ninja
greenlet
hpy
pydantic==1.10.13
readline==6.2.4.1
sqlparse==0.4.4
typing_extensions==4.8.0
django-ninja-extra==0.19.5
ninja-schema
djangorestframework
django-filter
jinja2
requests
supabase
python-dotenv
psycopg2-binary
djantic
icecream